requirements.txt
amqp==2.5.2
astor==0.8.1
autopep8==1.5.1
billiard==3.6.3.0
boto3==1.12.34
botocore==1.15.34
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
datasketch==1.5.1
dj-database-url==0.5.0
Django==2.2.10
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
kombu==4.6.8
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
